add methods printProductDetails and findProduct
The findProduct method has two diffrent methods with diffrent parameter options the first you can give a int and it will look for a product with that id or you can give it a name and it looks for a product with that name. If it finds the product it will then return it, if it doesnt find it then it will retun null.