added delivery method
This will incress the products quantity when you have it the id number of the product and the ammount you want to add. I also added a else to this so that if the id entered in the paramter is not found it will let the user know that so the are aware.